Crate google_classroom1

This documentation was generated from classroom crate version 2.0.0+20210323, where 20210323 is the exact revision of the classroom:v1 schema built by the mako code generator v2.0.0.

Everything else about the classroom v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

courses

aliases create, aliases delete, aliases list, announcements create, announcements delete, announcements get, announcements list, announcements modify assignees, announcements patch, course work create, course work delete, course work get, course work list, course work modify assignees, course work patch, course work student submissions get, course work student submissions list, course work student submissions modify attachments, course work student submissions patch, course work student submissions reclaim, course work student submissions return, course work student submissions turn in, course work materials create, course work materials delete, course work materials get, course work materials list, course work materials patch, create, delete, get, list, patch, students create, students delete, students get, students list, teachers create, teachers delete, teachers get, teachers list, topics create, topics delete, topics get, topics list, topics patch and update

invitations

accept, create, delete, get and list

registrations

create and delete

user profiles

get, guardian invitations create, guardian invitations get, guardian invitations list, guardian invitations patch, guardians delete, guardians get and guardians list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit().await

Or specifically ...

let r = hub.courses().aliases_create(...).doit().await
let r = hub.courses().aliases_delete(...).doit().await
let r = hub.courses().aliases_list(...).doit().await
let r = hub.courses().announcements_create(...).doit().await
let r = hub.courses().announcements_delete(...).doit().await
let r = hub.courses().announcements_get(...).doit().await
let r = hub.courses().announcements_list(...).doit().await
let r = hub.courses().announcements_modify_assignees(...).doit().await
let r = hub.courses().announcements_patch(...).doit().await
let r = hub.courses().course_work_student_submissions_get(...).doit().await
let r = hub.courses().course_work_student_submissions_list(...).doit().await
let r = hub.courses().course_work_student_submissions_modify_attachments(...).doit().await
let r = hub.courses().course_work_student_submissions_patch(...).doit().await
let r = hub.courses().course_work_student_submissions_reclaim(...).doit().await
let r = hub.courses().course_work_student_submissions_return(...).doit().await
let r = hub.courses().course_work_student_submissions_turn_in(...).doit().await
let r = hub.courses().course_work_create(...).doit().await
let r = hub.courses().course_work_delete(...).doit().await
let r = hub.courses().course_work_get(...).doit().await
let r = hub.courses().course_work_list(...).doit().await
let r = hub.courses().course_work_modify_assignees(...).doit().await
let r = hub.courses().course_work_patch(...).doit().await
let r = hub.courses().course_work_materials_create(...).doit().await
let r = hub.courses().course_work_materials_delete(...).doit().await
let r = hub.courses().course_work_materials_get(...).doit().await
let r = hub.courses().course_work_materials_list(...).doit().await
let r = hub.courses().course_work_materials_patch(...).doit().await
let r = hub.courses().students_create(...).doit().await
let r = hub.courses().students_delete(...).doit().await
let r = hub.courses().students_get(...).doit().await
let r = hub.courses().students_list(...).doit().await
let r = hub.courses().teachers_create(...).doit().await
let r = hub.courses().teachers_delete(...).doit().await
let r = hub.courses().teachers_get(...).doit().await
let r = hub.courses().teachers_list(...).doit().await
let r = hub.courses().topics_create(...).doit().await
let r = hub.courses().topics_delete(...).doit().await
let r = hub.courses().topics_get(...).doit().await
let r = hub.courses().topics_list(...).doit().await
let r = hub.courses().topics_patch(...).doit().await
let r = hub.courses().create(...).doit().await
let r = hub.courses().delete(...).doit().await
let r = hub.courses().get(...).doit().await
let r = hub.courses().list(...).doit().await
let r = hub.courses().patch(...).doit().await
let r = hub.courses().update(...).doit().await

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-classroom1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.14"
hyper-rustls = "^0.22"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^5.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_classroom1 as classroom1;
use classroom1::{Result, Error};
use std::default::Default;
use oauth2;
use classroom1::Classroom;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Classroom::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.courses().course_work_student_submissions_list("courseId", "courseWorkId")
             .user_id("takimata")
             .add_states("amet.")
             .page_token("duo")
             .page_size(-55)
             .late("gubergren")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be encodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Re-exports

pub use api::Classroom;

pub use client::Result;

pub use client::Error;

pub use client::Delegate;

Modules

api

client