Module google_classroom1::api

Structs

Announcement
Announcement created by a teacher for students of the course

Assignment
Additional details for assignments.

AssignmentSubmission
Student work for an assignment.

Attachment
Attachment added to student assignment work. When creating attachments, setting the form field is not supported.

Classroom
Central instance to access all Classroom related resource activities

CloudPubsubTopic
A reference to a Cloud Pub/Sub topic. To register for notifications, the owner of the topic must grant classroom-notifications@system.gserviceaccount.com the projects.topics.publish permission.

Course
A Course in Classroom.

CourseAlias
Alternative identifier for a course. An alias uniquely identifies a course. It must be unique within one of the following scopes: * domain: A domain-scoped alias is visible to all users within the alias creator's domain and can be created only by a domain admin. A domain-scoped alias is often used when a course has an identifier external to Classroom. * project: A project-scoped alias is visible to any request from an application using the Developer Console project ID that created the alias and can be created by any project. A project-scoped alias is often used when an application has alternative identifiers. A random value can also be used to avoid duplicate courses in the event of transmission failures, as retrying a request will return ALREADY_EXISTS if a previous one has succeeded.

CourseAliaseCreateCall
Creates an alias for a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create the alias or for access errors. * NOT_FOUND if the course does not exist. * ALREADY_EXISTS if the alias already exists. * FAILED_PRECONDITION if the alias requested does not make sense for the requesting user or course (for example, if a user not in a domain attempts to access a domain-scoped alias).

CourseAliaseDeleteCall
Deletes an alias of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to remove the alias or for access errors. * NOT_FOUND if the alias does not exist. * FAILED_PRECONDITION if the alias requested does not make sense for the requesting user or course (for example, if a user not in a domain attempts to delete a domain-scoped alias).

CourseAliaseListCall
Returns a list of aliases for a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the course or for access errors. * NOT_FOUND if the course does not exist.

CourseAnnouncementCreateCall
Creates an announcement. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course, create announcements in the requested course, share a Drive attachment, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist. * FAILED_PRECONDITION for the following request error: * AttachmentNotVisible

CourseAnnouncementDeleteCall
Deletes an announcement. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding announcement item. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding announcement, if the requesting user is not permitted to delete the requested course or for access errors. * FAILED_PRECONDITION if the requested announcement has already been deleted. * NOT_FOUND if no course exists with the requested ID.

CourseAnnouncementGetCall
Returns an announcement. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or announcement, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or announcement does not exist.

CourseAnnouncementListCall
Returns a list of announcements that the requester is permitted to view. Course students may only view PUBLISHED announcements. Course teachers and domain administrators may view all announcements. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist.

CourseAnnouncementModifyAssigneeCall
Modifies assignee mode and options of an announcement. Only a teacher of the course that contains the announcement may call this method. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or course work does not exist.

CourseAnnouncementPatchCall
Updates one or more fields of an announcement. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding announcement or for access errors. * INVALID_ARGUMENT if the request is malformed. * FAILED_PRECONDITION if the requested announcement has already been deleted. * NOT_FOUND if the requested course or announcement does not exist

CourseCourseWorkCreateCall
Creates course work. The resulting course work (and corresponding student submissions) are associated with the Developer Console project of the OAuth client ID used to make the request. Classroom API requests to modify course work and student submissions must be made with an OAuth client ID from the associated Developer Console project. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course, create course work in the requested course, share a Drive attachment, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist. * FAILED_PRECONDITION for the following request error: * AttachmentNotVisible

CourseCourseWorkDeleteCall
Deletes a course work. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding course work, if the requesting user is not permitted to delete the requested course or for access errors. * FAILED_PRECONDITION if the requested course work has already been deleted. * NOT_FOUND if no course exists with the requested ID.

CourseCourseWorkGetCall
Returns course work. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or course work does not exist.

CourseCourseWorkListCall
Returns a list of course work that the requester is permitted to view. Course students may only view PUBLISHED course work. Course teachers and domain administrators may view all course work. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist.

CourseCourseWorkMaterialCreateCall
Creates a course work material. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course, create course work material in the requested course, share a Drive attachment, or for access errors. * INVALID_ARGUMENT if the request is malformed or if more than 20 * materials are provided. * NOT_FOUND if the requested course does not exist. * FAILED_PRECONDITION for the following request error: * AttachmentNotVisible

CourseCourseWorkMaterialDeleteCall
Deletes a course work material. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work material item. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding course work material, if the requesting user is not permitted to delete the requested course or for access errors. * FAILED_PRECONDITION if the requested course work material has already been deleted. * NOT_FOUND if no course exists with the requested ID.

CourseCourseWorkMaterialGetCall
Returns a course work material. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work material, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or course work material does not exist.

CourseCourseWorkMaterialListCall
Returns a list of course work material that the requester is permitted to view. Course students may only view PUBLISHED course work material. Course teachers and domain administrators may view all course work material. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist.

CourseCourseWorkMaterialPatchCall
Updates one or more fields of a course work material. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project for access errors. * INVALID_ARGUMENT if the request is malformed. * FAILED_PRECONDITION if the requested course work material has already been deleted. * NOT_FOUND if the requested course or course work material does not exist

CourseCourseWorkModifyAssigneeCall
Modifies assignee mode and options of a coursework. Only a teacher of the course that contains the coursework may call this method. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or course work does not exist.

CourseCourseWorkPatchCall
Updates one or more fields of a course work. See google.classroom.v1.CourseWork for details of which fields may be updated and who may change them. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding course work, if the user is not permitted to make the requested modification to the student submission, or for access errors. * INVALID_ARGUMENT if the request is malformed. * FAILED_PRECONDITION if the requested course work has already been deleted. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCourseWorkStudentSubmissionGetCall
Returns a student submission. * PERMISSION_DENIED if the requesting user is not permitted to access the requested course, course work, or student submission or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCourseWorkStudentSubmissionListCall
Returns a list of student submissions that the requester is permitted to view, factoring in the OAuth scopes of the request. - may be specified as the course_work_id to include student submissions for multiple course work items. Course students may only view their own work. Course teachers and domain administrators may view all student submissions. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist.

CourseCourseWorkStudentSubmissionModifyAttachmentCall
Modifies attachments of student submission. Attachments may only be added to student submissions belonging to course work objects with a workType of ASSIGNMENT. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work, if the user is not permitted to modify attachments on the requested student submission, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCourseWorkStudentSubmissionPatchCall
Updates one or more fields of a student submission. See google.classroom.v1.StudentSubmission for details of which fields may be updated and who may change them. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding course work, if the user is not permitted to make the requested modification to the student submission, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCourseWorkStudentSubmissionReclaimCall
Reclaims a student submission on behalf of the student that owns it. Reclaiming a student submission transfers ownership of attached Drive files to the student and updates the submission state. Only the student that owns the requested student submission may call this method, and only for a student submission that has been turned in. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work, unsubmit the requested student submission, or for access errors. * FAILED_PRECONDITION if the student submission has not been turned in. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCourseWorkStudentSubmissionReturnCall
Returns a student submission. Returning a student submission transfers ownership of attached Drive files to the student and may also update the submission state. Unlike the Classroom application, returning a student submission does not set assignedGrade to the draftGrade value. Only a teacher of the course that contains the requested student submission may call this method. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work, return the requested student submission, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCourseWorkStudentSubmissionTurnInCall
Turns in a student submission. Turning in a student submission transfers ownership of attached Drive files to the teacher and may also update the submission state. This may only be called by the student that owns the specified student submission. This request must be made by the Developer Console project of the OAuth client ID used to create the corresponding course work item. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or course work, turn in the requested student submission, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course, course work, or student submission does not exist.

CourseCreateCall
Creates a course. The user specified in ownerId is the owner of the created course and added as a teacher. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create courses or for access errors. * NOT_FOUND if the primary teacher is not a valid user. * FAILED_PRECONDITION if the course owner's account is disabled or for the following request errors: * UserGroupsMembershipLimitReached * ALREADY_EXISTS if an alias was specified in the id and already exists.

CourseDeleteCall
Deletes a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID.

CourseGetCall
Returns a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID.

CourseListCall
Returns a list of courses that the requesting user is permitted to view, restricted to those that match the request. Returned courses are ordered by creation time, with the most recently created coming first. This method returns the following error codes: * PERMISSION_DENIED for access errors. * INVALID_ARGUMENT if the query argument is malformed. * NOT_FOUND if any users specified in the query arguments do not exist.

CourseMaterial
A material attached to a course as part of a material set.

CourseMaterialSet
A set of materials that appears on the "About" page of the course. These materials might include a syllabus, schedule, or other background information relating to the course as a whole.

CourseMethods
A builder providing access to all methods supported on course resources. It is not used directly, but through the Classroom hub.

CoursePatchCall
Updates one or more fields in a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to modify the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID. * INVALID_ARGUMENT if invalid fields are specified in the update mask or if no update mask is supplied. * FAILED_PRECONDITION for the following request errors: * CourseNotModifiable

CourseRosterChangesInfo
Information about a Feed with a feed_type of COURSE_ROSTER_CHANGES.

CourseStudentCreateCall
Adds a user as a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create students in this course or for access errors. * NOT_FOUND if the requested course ID does not exist. * FAILED_PRECONDITION if the requested user's account is disabled, for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * UserGroupsMembershipLimitReached * ALREADY_EXISTS if the user is already a student or teacher in the course.

CourseStudentDeleteCall
Deletes a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete students of this course or for access errors. * NOT_FOUND if no student of this course has the requested ID or if the course does not exist.

CourseStudentGetCall
Returns a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view students of this course or for access errors. * NOT_FOUND if no student of this course has the requested ID or if the course does not exist.

CourseStudentListCall
Returns a list of students of this course that the requester is permitted to view. This method returns the following error codes: * NOT_FOUND if the course does not exist. * PERMISSION_DENIED for access errors.

CourseTeacherCreateCall
Creates a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create teachers in this course or for access errors. * NOT_FOUND if the requested course ID does not exist. * FAILED_PRECONDITION if the requested user's account is disabled, for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * CourseTeacherLimitReached * UserGroupsMembershipLimitReached * ALREADY_EXISTS if the user is already a teacher or student in the course.

CourseTeacherDeleteCall
Deletes a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete teachers of this course or for access errors. * NOT_FOUND if no teacher of this course has the requested ID or if the course does not exist. * FAILED_PRECONDITION if the requested ID belongs to the primary teacher of this course.

CourseTeacherGetCall
Returns a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view teachers of this course or for access errors. * NOT_FOUND if no teacher of this course has the requested ID or if the course does not exist.

CourseTeacherListCall
Returns a list of teachers of this course that the requester is permitted to view. This method returns the following error codes: * NOT_FOUND if the course does not exist. * PERMISSION_DENIED for access errors.

CourseTopicCreateCall
Creates a topic. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course, create a topic in the requested course, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist.

CourseTopicDeleteCall
Deletes a topic. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not allowed to delete the requested topic or for access errors. * FAILED_PRECONDITION if the requested topic has already been deleted. * NOT_FOUND if no course or topic exists with the requested ID.

CourseTopicGetCall
Returns a topic. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or topic, or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or topic does not exist.

CourseTopicListCall
Returns the list of topics that the requester is permitted to view. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course does not exist.

CourseTopicPatchCall
Updates one or more fields of a topic. This method returns the following error codes: * PERMISSION_DENIED if the requesting developer project did not create the corresponding topic or for access errors. * INVALID_ARGUMENT if the request is malformed. * NOT_FOUND if the requested course or topic does not exist

CourseUpdateCall
Updates a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to modify the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID. * FAILED_PRECONDITION for the following request errors: * CourseNotModifiable

CourseWork
Course work created by a teacher for students of the course.

CourseWorkChangesInfo
Information about a Feed with a feed_type of COURSE_WORK_CHANGES.

CourseWorkMaterial
Course work material created by a teacher for students of the course

Date
Represents a whole or partial calendar date, such as a birthday. The time of day and time zone are either specified elsewhere or are insignificant. The date is relative to the Gregorian Calendar. This can represent one of the following: * A full date, with non-zero year, month, and day values * A month and day value, with a zero year, such as an anniversary * A year on its own, with zero month and day values * A year and month value, with a zero day, such as a credit card expiration date Related types are google.type.TimeOfDay and google.protobuf.Timestamp.

DriveFile
Representation of a Google Drive file.

DriveFolder
Representation of a Google Drive folder.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

Feed
A class of notifications that an application can register to receive. For example: "all roster changes for a domain".

Form
Google Forms item.

GlobalPermission
Global user permission description.

GradeHistory
The history of each grade on this submission.

Guardian
Association between a student and a guardian of that student. The guardian may receive information about the student's course work.

GuardianInvitation
An invitation to become the guardian of a specified user, sent to a specified email address.

IndividualStudentsOptions
Assignee details about a coursework/announcement. This field is set if and only if assigneeMode is INDIVIDUAL_STUDENTS.

Invitation
An invitation to join a course.

InvitationAcceptCall
Accepts an invitation, removing it and adding the invited user to the teachers or students (as appropriate) of the specified course. Only the invited user may accept an invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to accept the requested invitation or for access errors. * FAILED_PRECONDITION for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * CourseTeacherLimitReached * UserGroupsMembershipLimitReached * NOT_FOUND if no invitation exists with the requested ID.

InvitationCreateCall
Creates an invitation. Only one invitation for a user and course may exist at a time. Delete and re-create an invitation to make changes. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create invitations for this course or for access errors. * NOT_FOUND if the course or the user does not exist. * FAILED_PRECONDITION if the requested user's account is disabled or if the user already has this role or a role with greater permissions. * ALREADY_EXISTS if an invitation for the specified user and course already exists.

InvitationDeleteCall
Deletes an invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete the requested invitation or for access errors. * NOT_FOUND if no invitation exists with the requested ID.

InvitationGetCall
Returns an invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view the requested invitation or for access errors. * NOT_FOUND if no invitation exists with the requested ID.

InvitationListCall
Returns a list of invitations that the requesting user is permitted to view, restricted to those that match the list request. Note: At least one of user_id or course_id must be supplied. Both fields can be supplied. This method returns the following error codes: * PERMISSION_DENIED for access errors.

InvitationMethods
A builder providing access to all methods supported on invitation resources. It is not used directly, but through the Classroom hub.

Link
URL item.

ListAnnouncementsResponse
Response when listing course work.

ListCourseAliasesResponse
Response when listing course aliases.

ListCourseWorkMaterialResponse
Response when listing course work material.

ListCourseWorkResponse
Response when listing course work.

ListCoursesResponse
Response when listing courses.

ListGuardianInvitationsResponse
Response when listing guardian invitations.

ListGuardiansResponse
Response when listing guardians.

ListInvitationsResponse
Response when listing invitations.

ListStudentSubmissionsResponse
Response when listing student submissions.

ListStudentsResponse
Response when listing students.

ListTeachersResponse
Response when listing teachers.

ListTopicResponse
Response when listing topics.

Material
Material attached to course work. When creating attachments, setting the form field is not supported.

ModifyAnnouncementAssigneesRequest
Request to modify assignee mode and options of an announcement.

ModifyAttachmentsRequest
Request to modify the attachments of a student submission.

ModifyCourseWorkAssigneesRequest
Request to modify assignee mode and options of a coursework.

ModifyIndividualStudentsOptions
Contains fields to add or remove students from a course work or announcement where the assigneeMode is set to INDIVIDUAL_STUDENTS.

MultipleChoiceQuestion
Additional details for multiple-choice questions.

MultipleChoiceSubmission
Student work for a multiple-choice question.

Name
Details of the user's name.

ReclaimStudentSubmissionRequest
Request to reclaim a student submission.

Registration
An instruction to Classroom to send notifications from the feed to the provided destination.

RegistrationCreateCall
Creates a Registration, causing Classroom to start sending notifications from the provided feed to the destination provided in cloudPubSubTopic. Returns the created Registration. Currently, this will be the same as the argument, but with server-assigned fields such as expiry_time and id filled in. Note that any value specified for the expiry_time or id fields will be ignored. While Classroom may validate the cloudPubSubTopic and return errors on a best effort basis, it is the caller's responsibility to ensure that it exists and that Classroom has permission to publish to it. This method may return the following error codes: * PERMISSION_DENIED if: * the authenticated user does not have permission to receive notifications from the requested field; or * the current user has not granted access to the current Cloud project with the appropriate scope for the requested feed. Note that domain-wide delegation of authority is not currently supported for this purpose. If the request has the appropriate scope, but no grant exists, a Request Errors is returned. * another access error is encountered. * INVALID_ARGUMENT if: * no cloudPubsubTopic is specified, or the specified cloudPubsubTopic is not valid; or * no feed is specified, or the specified feed is not valid. * NOT_FOUND if: * the specified feed cannot be located, or the requesting user does not have permission to determine whether or not it exists; or * the specified cloudPubsubTopic cannot be located, or Classroom has not been granted permission to publish to it.

RegistrationDeleteCall
Deletes a Registration, causing Classroom to stop sending notifications for that Registration.

RegistrationMethods
A builder providing access to all methods supported on registration resources. It is not used directly, but through the Classroom hub.

ReturnStudentSubmissionRequest
Request to return a student submission.

SharedDriveFile
Drive file that is used as material for course work.

ShortAnswerSubmission
Student work for a short answer question.

StateHistory
The history of each state this submission has been in.

Student
Student in a course.

StudentSubmission
Student submission for course work. StudentSubmission items are generated when a CourseWork item is created. StudentSubmissions that have never been accessed (i.e. with state = NEW) may not have a creation time or update time.

SubmissionHistory
The history of the submission. This currently includes state and grade histories.

Teacher
Teacher of a course.

TimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

Topic
Topic created by a teacher for the course

TurnInStudentSubmissionRequest
Request to turn in a student submission.

UserProfile
Global information for a user.

UserProfileGetCall
Returns a user profile. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access this user profile, if no profile exists with the requested ID, or for access errors.

UserProfileGuardianDeleteCall
Deletes a guardian. The guardian will no longer receive guardian notifications and the guardian will no longer be accessible via the API. This method returns the following error codes: * PERMISSION_DENIED if no user that matches the provided student_id is visible to the requesting user, if the requesting user is not permitted to manage guardians for the student identified by the student_id, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API). * NOT_FOUND if the requesting user is permitted to modify guardians for the requested student_id, but no Guardian record exists for that student with the provided guardian_id.

UserProfileGuardianGetCall
Returns a specific guardian. This method returns the following error codes: * PERMISSION_DENIED if no user that matches the provided student_id is visible to the requesting user, if the requesting user is not permitted to view guardian information for the student identified by the student_id, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). * NOT_FOUND if the requesting user is permitted to view guardians for the requested student_id, but no Guardian record exists for that student that matches the provided guardian_id.

UserProfileGuardianInvitationCreateCall
Creates a guardian invitation, and sends an email to the guardian asking them to confirm that they are the student's guardian. Once the guardian accepts the invitation, their state will change to COMPLETED and they will start receiving guardian notifications. A Guardian resource will also be created to represent the active guardian. The request object must have the student_id and invited_email_address fields set. Failing to set these fields, or setting any other fields in the request, will result in an error. This method returns the following error codes: * PERMISSION_DENIED if the current user does not have permission to manage guardians, if the guardian in question has already rejected too many requests for that student, if guardians are not enabled for the domain in question, or for other access errors. * RESOURCE_EXHAUSTED if the student or guardian has exceeded the guardian link limit. * INVALID_ARGUMENT if the guardian email address is not valid (for example, if it is too long), or if the format of the student ID provided cannot be recognized (it is not an email address, nor a user_id from this API). This error will also be returned if read-only fields are set, or if the state field is set to to a value other than PENDING. * NOT_FOUND if the student ID provided is a valid student ID, but Classroom has no record of that student. * ALREADY_EXISTS if there is already a pending guardian invitation for the student and invited_email_address provided, or if the provided invited_email_address matches the Google account of an existing Guardian for this user.

UserProfileGuardianInvitationGetCall
Returns a specific guardian invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view guardian invitations for the student identified by the student_id, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). * NOT_FOUND if Classroom cannot find any record of the given student or invitation_id. May also be returned if the student exists, but the requesting user does not have access to see that student.

UserProfileGuardianInvitationListCall
Returns a list of guardian invitations that the requesting user is permitted to view, filtered by the parameters provided. This method returns the following error codes: * PERMISSION_DENIED if a student_id is specified, and the requesting user is not permitted to view guardian invitations for that student, if "-" is specified as the student_id and the user is not a domain administrator, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). May also be returned if an invalid page_token or state is provided. * NOT_FOUND if a student_id is specified, and its format can be recognized, but Classroom has no record of that student.

UserProfileGuardianInvitationPatchCall
Modifies a guardian invitation. Currently, the only valid modification is to change the state from PENDING to COMPLETE. This has the effect of withdrawing the invitation. This method returns the following error codes: * PERMISSION_DENIED if the current user does not have permission to manage guardians, if guardians are not enabled for the domain in question or for other access errors. * FAILED_PRECONDITION if the guardian link is not in the PENDING state. * INVALID_ARGUMENT if the format of the student ID provided cannot be recognized (it is not an email address, nor a user_id from this API), or if the passed GuardianInvitation has a state other than COMPLETE, or if it modifies fields other than state. * NOT_FOUND if the student ID provided is a valid student ID, but Classroom has no record of that student, or if the id field does not refer to a guardian invitation known to Classroom.

UserProfileGuardianListCall
Returns a list of guardians that the requesting user is permitted to view, restricted to those that match the request. To list guardians for any student that the requesting user may view guardians for, use the literal character - for the student ID. This method returns the following error codes: * PERMISSION_DENIED if a student_id is specified, and the requesting user is not permitted to view guardian information for that student, if "-" is specified as the student_id and the user is not a domain administrator, if guardians are not enabled for the domain in question, if the invited_email_address filter is set by a user who is not a domain administrator, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). May also be returned if an invalid page_token is provided. * NOT_FOUND if a student_id is specified, and its format can be recognized, but Classroom has no record of that student.

UserProfileMethods
A builder providing access to all methods supported on userProfile resources. It is not used directly, but through the Classroom hub.

YouTubeVideo
YouTube video item.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.