Struct google_classroom1::Course

pub struct Course {
    pub update_time: Option<String>,
    pub description: Option<String>,
    pub alternate_link: Option<String>,
    pub enrollment_code: Option<String>,
    pub section: Option<String>,
    pub creation_time: Option<String>,
    pub room: Option<String>,
    pub owner_id: Option<String>,
    pub course_state: Option<String>,
    pub id: Option<String>,
    pub description_heading: Option<String>,
    pub name: Option<String>,
}

A Course in Classroom.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update courses (request|response)

list courses (none)

teachers list courses (none)

get courses (response)

students delete courses (none)

patch courses (request|response)

aliases create courses (none)

teachers delete courses (none)

students create courses (none)

students get courses (none)

aliases delete courses (none)

teachers create courses (none)

teachers get courses (none)

create courses (request|response)

students list courses (none)

aliases list courses (none)

delete courses (none)

Fields

update_time

Time of the most recent update to this course. Specifying this field in a course update mask will result in an error. Read-only.

description

Optional description. For example, "We'll be learning about the structure of living creatures from a combination of textbooks, guest lectures, and lab work. Expect to be excited!" If set, this field must be a valid UTF-8 string and no longer than 30,000 characters.

alternate_link

Absolute link to this course in the Classroom web UI. Read-only.

enrollment_code

Enrollment code to use when joining this course. Specifying this field in a course update mask will result in an error. Read-only.

section

Section of the course. For example, "Period 2". If set, this field must be a valid UTF-8 string and no longer than 2800 characters.

creation_time

Creation time of the course. Specifying this field in a course update mask will result in an error. Read-only.

room

Optional room location. For example, "301". If set, this field must be a valid UTF-8 string and no longer than 650 characters.

owner_id

The identifier of the owner of a course. When specified as a parameter of a create course request, this field is required. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user This must be set in a create request. Specifying this field in a course update mask will result in an INVALID_ARGUMENT error.

course_state

State of the course. If unspecified, the default state is PROVISIONED.

id

Identifier for this course assigned by Classroom. When creating a course, you may optionally set this identifier to an alias string in the request to create a corresponding alias. The id is still assigned by Classroom and cannot be updated after the course is created. Specifying this field in a course update mask will result in an error.

description_heading

Optional heading for the description. For example, "Welcome to 10th Grade Biology." If set, this field must be a valid UTF-8 string and no longer than 3600 characters.

name

Name of the course. For example, "10th Grade Biology". The name is required. It must be between 1 and 750 characters and a valid UTF-8 string.