Struct google_classroom1::StudentSubmission

pub struct StudentSubmission {
    pub update_time: Option<String>,
    pub course_id: Option<String>,
    pub user_id: Option<String>,
    pub associated_with_developer: Option<bool>,
    pub course_work_type: Option<String>,
    pub assigned_grade: Option<f64>,
    pub id: Option<String>,
    pub draft_grade: Option<f64>,
    pub alternate_link: Option<String>,
    pub multiple_choice_submission: Option<MultipleChoiceSubmission>,
    pub creation_time: Option<String>,
    pub short_answer_submission: Option<ShortAnswerSubmission>,
    pub late: Option<bool>,
    pub state: Option<String>,
    pub course_work_id: Option<String>,
    pub assignment_submission: Option<AssignmentSubmission>,
}

Student submission for course work. StudentSubmission items are generated when a CourseWork item is created. StudentSubmissions that have never been accessed (i.e. with state = NEW) may not have a creation time or update time.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

course work student submissions patch courses (request|response)

course work student submissions get courses (response)

course work student submissions modify attachments courses (response)

Fields
update_time: Option<String>

Last update time of this submission. This may be unset if the student has not accessed this item. Read-only.

course_id: Option<String>

Identifier of the course. Read-only.

user_id: Option<String>

Identifier for the student that owns this submission. Read-only.

associated_with_developer: Option<bool>

Whether this student submission is associated with the Developer Console project making the request. See google.classroom.Work.CreateCourseWork for more details. Read-only.

course_work_type: Option<String>

Type of course work this submission is for. Read-only.

assigned_grade: Option<f64>

Optional grade. If unset, no grade was set. This must be a non-negative integer value. This may be modified only by course teachers.

id: Option<String>

Classroom-assigned Identifier for the student submission. This is unique among submissions for the relevant course work. Read-only.

draft_grade: Option<f64>

Optional pending grade. If unset, no grade was set. This must be a non-negative integer value. This is only visible to and modifiable by course teachers.

alternate_link: Option<String>

Absolute link to the submission in the Classroom web UI. Read-only.

multiple_choice_submission: Option<MultipleChoiceSubmission>

Submission content when course_work_type is MULTIPLE_CHOICE_QUESTION.

creation_time: Option<String>

Creation time of this submission. This may be unset if the student has not accessed this item. Read-only.

short_answer_submission: Option<ShortAnswerSubmission>

Submission content when course_work_type is SHORT_ANSWER_QUESTION.

late: Option<bool>

Whether this submission is late. Read-only.

state: Option<String>

State of this submission. Read-only.

course_work_id: Option<String>

Identifier for the course work this corresponds to. Read-only.

assignment_submission: Option<AssignmentSubmission>

Submission content when course_work_type is ASSIGNMENT .