Struct google_classroom1::UserProfileMethods

pub struct UserProfileMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on userProfile resources. It is not used directly, but through the Classroom hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_classroom1 as classroom1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use classroom1::Classroom;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Classroom::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `guardian_invitations_create(...)`, `guardian_invitations_get(...)`, `guardian_invitations_list(...)`, `guardian_invitations_patch(...)`, `guardians_delete(...)`, `guardians_get(...)` and `guardians_list(...)`
// to build up your call.
let rb = hub.user_profiles();

Methods

impl<'a, C, A> UserProfileMethods<'a, C, A>

fn guardians_get(&self, student_id: &str, guardian_id: &str) -> UserProfileGuardianGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a specific guardian. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view guardian information for the student identified by the student_id, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). * NOT_FOUND if Classroom cannot find any record of the given student or guardian_id, or if the guardian has been disabled.

Arguments

studentId - The student whose guardian is being requested. One of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

guardianId - The id field from a Guardian.

fn guardian_invitations_create(&self, request: GuardianInvitation, student_id: &str) -> UserProfileGuardianInvitationCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a guardian invitation, and sends an email to the guardian asking them to confirm that they are the student's guardian. Once the guardian accepts the invitation, their state will change to COMPLETED and they will start receiving guardian notifications. A Guardian resource will also be created to represent the active guardian. The request object must have the student_id and invited_email_address fields set. Failing to set these fields, or setting any other fields in the request, will result in an error. This method returns the following error codes: * PERMISSION_DENIED if the current user does not have permission to manage guardians, if the guardian in question has already rejected too many requests for that student, if guardians are not enabled for the domain in question, or for other access errors. * RESOURCE_EXHAUSTED if the student or guardian has exceeded the guardian link limit. * INVALID_ARGUMENT if the guardian email address is not valid (for example, if it is too long), or if the format of the student ID provided cannot be recognized (it is not an email address, nor a user_id from this API). This error will also be returned if read-only fields are set, or if the state field is set to to a value other than PENDING. * NOT_FOUND if the student ID provided is a valid student ID, but Classroom has no record of that student. * ALREADY_EXISTS if there is already a pending guardian invitation for the student and invited_email_address provided, or if the provided invited_email_address matches the Google account of an existing Guardian for this user.

Arguments

request - No description provided.

studentId - ID of the student (in standard format)

fn get(&self, user_id: &str) -> UserProfileGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a user profile. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access this user profile or if no profile exists with the requested ID or for access errors.

Arguments

userId - Identifier of the profile to return. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn guardian_invitations_patch(&self, request: GuardianInvitation, student_id: &str, invitation_id: &str) -> UserProfileGuardianInvitationPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies a guardian invitation. Currently, the only valid modification is to change the state from PENDING to COMPLETE. This has the effect of withdrawing the invitation. This method returns the following error codes: * PERMISSION_DENIED if the current user does not have permission to manage guardians, if guardians are not enabled for the domain in question or for other access errors. * FAILED_PRECONDITION if the guardian link is not in the PENDING state. * INVALID_ARGUMENT if the format of the student ID provided cannot be recognized (it is not an email address, nor a user_id from this API), or if the passed GuardianInvitation has a state other than COMPLETE, or if it modifies fields other than state. * NOT_FOUND if the student ID provided is a valid student ID, but Classroom has no record of that student, or if the id field does not refer to a guardian invitation known to Classroom.

Arguments

request - No description provided.

studentId - The ID of the student whose guardian invitation is to be modified.

invitationId - The id field of the GuardianInvitation to be modified.

fn guardian_invitations_get(&self, student_id: &str, invitation_id: &str) -> UserProfileGuardianInvitationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a specific guardian invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view guardian invitations for the student identified by the student_id, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). * NOT_FOUND if Classroom cannot find any record of the given student or invitation_id. May also be returned if the student exists, but the requesting user does not have access to see that student.

Arguments

studentId - The ID of the student whose guardian invitation is being requested.

invitationId - The id field of the GuardianInvitation being requested.

fn guardians_list(&self, student_id: &str) -> UserProfileGuardianListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of guardians that the requesting user is permitted to view, restricted to those that match the request. This method returns the following error codes: * PERMISSION_DENIED if a student_id is specified, and the requesting user is not permitted to view guardian information for that student, if guardians are not enabled for the domain in question, if the invited_email_address filter is set by a user who is not a domain administrator, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). May also be returned if an invalid page_token is provided. * NOT_FOUND if a student_id is specified, and its format can be recognized, but Classroom has no record of that student.

Arguments

studentId - Filter results by the student who the guardian is linked to. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn guardian_invitations_list(&self, student_id: &str) -> UserProfileGuardianInvitationListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of guardian invitations that the requesting user is permitted to view, filtered by the parameters provided. This method returns the following error codes: * PERMISSION_DENIED if a student_id is specified, and the requesting user is not permitted to view guardian invitations for that student, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API, nor the literal string me). May also be returned if an invalid page_token or state is provided. * NOT_FOUND if a student_id is specified, and its format can be recognized, but Classroom has no record of that student.

Arguments

studentId - The ID of the student whose guardian invitations are to be returned. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn guardians_delete(&self, student_id: &str, guardian_id: &str) -> UserProfileGuardianDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a guardian. The guardian will no longer receive guardian notifications and the guardian will no longer be accessible via the API. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to manage guardians for the student identified by the student_id, if guardians are not enabled for the domain in question, or for other access errors. * INVALID_ARGUMENT if a student_id is specified, but its format cannot be recognized (it is not an email address, nor a student_id from the API). * NOT_FOUND if Classroom cannot find any record of the given student_id or guardian_id, or if the guardian has already been disabled.

Arguments

studentId - The student whose guardian is to be deleted. One of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

guardianId - The id field from a Guardian.