Crate google_classroom1

This documentation was generated from classroom crate version 0.1.14+20151013, where 20151013 is the exact revision of the classroom:v1 schema built by the mako code generator v0.1.14.

Everything else about the classroom v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

courses

aliases create, aliases delete, aliases list, create, delete, get, list, patch, students create, students delete, students get, students list, teachers create, teachers delete, teachers get, teachers list and update

invitations

accept, create, delete, get and list

user profiles

get

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.courses().update(...).doit()
let r = hub.courses().list(...).doit()
let r = hub.courses().teachers_list(...).doit()
let r = hub.courses().get(...).doit()
let r = hub.courses().students_delete(...).doit()
let r = hub.courses().patch(...).doit()
let r = hub.courses().aliases_create(...).doit()
let r = hub.courses().teachers_delete(...).doit()
let r = hub.courses().students_create(...).doit()
let r = hub.courses().students_get(...).doit()
let r = hub.courses().aliases_delete(...).doit()
let r = hub.courses().teachers_create(...).doit()
let r = hub.courses().teachers_get(...).doit()
let r = hub.courses().create(...).doit()
let r = hub.courses().students_list(...).doit()
let r = hub.courses().aliases_list(...).doit()
let r = hub.courses().delete(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-classroom1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_classroom1 as classroom1;
use classroom1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use classroom1::Classroom;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Classroom::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.courses().list()
             .teacher_id("sea")
             .student_id("et")
             .page_token("duo")
             .page_size(-21)
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Classroom
Central instance to access all Classroom related resource activities

Course
A Course in Classroom.

CourseAlias
Alternative identifier for a course. An alias uniquely identifies a course. It must be unique within one of the following scopes: * domain: A domain-scoped alias is visible to all users within the alias creator's domain and can be created only by a domain admin. A domain-scoped alias is often used when a course has an identifier external to Classroom. * project: A project-scoped alias is visible to any request from an application using the Developer Console project ID that created the alias and can be created by any project. A project-scoped alias is often used when an application has alternative identifiers. A random value can also be used to avoid duplicate courses in the event of transmission failures, as retrying a request will return ALREADY_EXISTS if a previous one has succeeded.

CourseAliaseCreateCall
Creates an alias for a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create the alias or for access errors. * NOT_FOUND if the course does not exist. * ALREADY_EXISTS if the alias already exists.

CourseAliaseDeleteCall
Deletes an alias of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to remove the alias or for access errors. * NOT_FOUND if the alias does not exist.

CourseAliaseListCall
Returns a list of aliases for a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the course or for access errors. * NOT_FOUND if the course does not exist.

CourseCreateCall
Creates a course. The user specified in ownerId is the owner of the created course and added as a teacher. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create courses or for access errors. * NOT_FOUND if the primary teacher is not a valid user. * FAILED_PRECONDITION if the course owner's account is disabled or for the following request errors: * UserGroupsMembershipLimitReached * ALREADY_EXISTS if an alias was specified in the id and already exists.

CourseDeleteCall
Deletes a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID.

CourseGetCall
Returns a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID.

CourseListCall
Returns a list of courses that the requesting user is permitted to view, restricted to those that match the request. This method returns the following error codes: * PERMISSION_DENIED for access errors. * INVALID_ARGUMENT if the query argument is malformed. * NOT_FOUND if any users specified in the query arguments do not exist.

CourseMethods
A builder providing access to all methods supported on course resources. It is not used directly, but through the Classroom hub.

CoursePatchCall
Updates one or more fields in a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to modify the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID. * INVALID_ARGUMENT if invalid fields are specified in the update mask or if no update mask is supplied. * FAILED_PRECONDITION for the following request errors: * CourseNotModifiable

CourseStudentCreateCall
Adds a user as a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create students in this course or for access errors. * NOT_FOUND if the requested course ID does not exist. * FAILED_PRECONDITION if the requested user's account is disabled, for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * UserGroupsMembershipLimitReached * ALREADY_EXISTS if the user is already a student or teacher in the course.

CourseStudentDeleteCall
Deletes a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete students of this course or for access errors. * NOT_FOUND if no student of this course has the requested ID or if the course does not exist.

CourseStudentGetCall
Returns a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view students of this course or for access errors. * NOT_FOUND if no student of this course has the requested ID or if the course does not exist.

CourseStudentListCall
Returns a list of students of this course that the requester is permitted to view. This method returns the following error codes: * NOT_FOUND if the course does not exist. * PERMISSION_DENIED for access errors.

CourseTeacherCreateCall
Creates a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create teachers in this course or for access errors. * NOT_FOUND if the requested course ID does not exist. * FAILED_PRECONDITION if the requested user's account is disabled, for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * CourseTeacherLimitReached * UserGroupsMembershipLimitReached * ALREADY_EXISTS if the user is already a teacher or student in the course.

CourseTeacherDeleteCall
Deletes a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete teachers of this course or for access errors. * NOT_FOUND if no teacher of this course has the requested ID or if the course does not exist. * FAILED_PRECONDITION if the requested ID belongs to the primary teacher of this course.

CourseTeacherGetCall
Returns a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view teachers of this course or for access errors. * NOT_FOUND if no teacher of this course has the requested ID or if the course does not exist.

CourseTeacherListCall
Returns a list of teachers of this course that the requester is permitted to view. This method returns the following error codes: * NOT_FOUND if the course does not exist. * PERMISSION_DENIED for access errors.

CourseUpdateCall
Updates a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to modify the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID. * FAILED_PRECONDITION for the following request errors: * CourseNotModifiable

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GlobalPermission
Global user permission description.

Invitation
An invitation to join a course.

InvitationAcceptCall
Accepts an invitation, removing it and adding the invited user to the teachers or students (as appropriate) of the specified course. Only the invited user may accept an invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to accept the requested invitation or for access errors. * FAILED_PRECONDITION for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * CourseTeacherLimitReached * UserGroupsMembershipLimitReached * NOT_FOUND if no invitation exists with the requested ID.

InvitationCreateCall
Creates an invitation. Only one invitation for a user and course may exist at a time. Delete and re-create an invitation to make changes. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create invitations for this course or for access errors. * NOT_FOUND if the course or the user does not exist. * FAILED_PRECONDITION if the requested user's account is disabled or if the user already has this role or a role with greater permissions. * ALREADY_EXISTS if an invitation for the specified user and course already exists.

InvitationDeleteCall
Deletes an invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete the requested invitation or for access errors. * NOT_FOUND if no invitation exists with the requested ID.

InvitationGetCall
Returns an invitation. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view the requested invitation or for access errors. * NOT_FOUND if no invitation exists with the requested ID.

InvitationListCall
Returns a list of invitations that the requesting user is permitted to view, restricted to those that match the list request. Note: At least one of user_id or course_id must be supplied. Both fields can be supplied. This method returns the following error codes: * PERMISSION_DENIED for access errors.

InvitationMethods
A builder providing access to all methods supported on invitation resources. It is not used directly, but through the Classroom hub.

ListCourseAliasesResponse
Response when listing course aliases.

ListCoursesResponse
Response when listing courses.

ListInvitationsResponse
Response when listing invitations.

ListStudentsResponse
Response when listing students.

ListTeachersResponse
Response when listing teachers.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Name
Details of the user's name.

Student
Student in a course.

Teacher
Teacher of a course.

UserProfile
Global information for a user.

UserProfileGetCall
Returns a user profile. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access this user profile or if no profile exists with the requested ID or for access errors.

UserProfileMethods
A builder providing access to all methods supported on userProfile resources. It is not used directly, but through the Classroom hub.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.