Crate google_classroom1

This documentation was generated from classroom crate version 1.0.9+20190701, where 20190701 is the exact revision of the classroom:v1 schema built by the mako code generator v1.0.9.

Everything else about the classroom v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

courses

aliases create, aliases delete, aliases list, announcements create, announcements delete, announcements get, announcements list, announcements modify assignees, announcements patch, course work create, course work delete, course work get, course work list, course work modify assignees, course work patch, course work student submissions get, course work student submissions list, course work student submissions modify attachments, course work student submissions patch, course work student submissions reclaim, course work student submissions return, course work student submissions turn in, create, delete, get, list, patch, students create, students delete, students get, students list, teachers create, teachers delete, teachers get, teachers list, topics create, topics delete, topics get, topics list, topics patch and update

invitations

accept, create, delete, get and list

registrations

create and delete

user profiles

get, guardian invitations create, guardian invitations get, guardian invitations list, guardian invitations patch, guardians delete, guardians get and guardians list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.courses().announcements_modify_assignees(...).doit()
let r = hub.courses().course_work_student_submissions_patch(...).doit()
let r = hub.courses().announcements_list(...).doit()
let r = hub.courses().teachers_list(...).doit()
let r = hub.courses().get(...).doit()
let r = hub.courses().update(...).doit()
let r = hub.courses().students_get(...).doit()
let r = hub.courses().teachers_get(...).doit()
let r = hub.courses().course_work_list(...).doit()
let r = hub.courses().topics_list(...).doit()
let r = hub.courses().course_work_student_submissions_reclaim(...).doit()
let r = hub.courses().course_work_student_submissions_list(...).doit()
let r = hub.courses().course_work_student_submissions_turn_in(...).doit()
let r = hub.courses().course_work_student_submissions_modify_attachments(...).doit()
let r = hub.courses().announcements_get(...).doit()
let r = hub.courses().course_work_student_submissions_return(...).doit()
let r = hub.courses().aliases_list(...).doit()
let r = hub.courses().course_work_get(...).doit()
let r = hub.courses().course_work_create(...).doit()
let r = hub.courses().list(...).doit()
let r = hub.courses().announcements_create(...).doit()
let r = hub.courses().topics_patch(...).doit()
let r = hub.courses().announcements_patch(...).doit()
let r = hub.courses().aliases_create(...).doit()
let r = hub.courses().students_create(...).doit()
let r = hub.courses().course_work_modify_assignees(...).doit()
let r = hub.courses().aliases_delete(...).doit()
let r = hub.courses().course_work_delete(...).doit()
let r = hub.courses().create(...).doit()
let r = hub.courses().students_list(...).doit()
let r = hub.courses().delete(...).doit()
let r = hub.courses().course_work_patch(...).doit()
let r = hub.courses().topics_delete(...).doit()
let r = hub.courses().patch(...).doit()
let r = hub.courses().students_delete(...).doit()
let r = hub.courses().teachers_delete(...).doit()
let r = hub.courses().teachers_create(...).doit()
let r = hub.courses().course_work_student_submissions_get(...).doit()
let r = hub.courses().topics_get(...).doit()
let r = hub.courses().announcements_delete(...).doit()
let r = hub.courses().topics_create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-classroom1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_classroom1 as classroom1;
use classroom1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use classroom1::Classroom;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Classroom::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.courses().course_work_student_submissions_list("courseId", "courseWorkId")
             .user_id("justo")
             .add_states("amet.")
             .page_token("erat")
             .page_size(-35)
             .late("sea")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Announcement
Announcement created by a teacher for students of the course

Assignment
Additional details for assignments.

AssignmentSubmission
Student work for an assignment.

Attachment
Attachment added to student assignment work.

Chunk

Classroom
Central instance to access all Classroom related resource activities

CloudPubsubTopic
A reference to a Cloud Pub/Sub topic.

ContentRange
Implements the Content-Range header, for serialization only

Course
A Course in Classroom.

CourseAlias
Alternative identifier for a course.

CourseAliaseCreateCall
Creates an alias for a course.

CourseAliaseDeleteCall
Deletes an alias of a course.

CourseAliaseListCall
Returns a list of aliases for a course.

CourseAnnouncementCreateCall
Creates an announcement.

CourseAnnouncementDeleteCall
Deletes an announcement.

CourseAnnouncementGetCall
Returns an announcement.

CourseAnnouncementListCall
Returns a list of announcements that the requester is permitted to view.

CourseAnnouncementModifyAssigneeCall
Modifies assignee mode and options of an announcement.

CourseAnnouncementPatchCall
Updates one or more fields of an announcement.

CourseCourseWorkCreateCall
Creates course work.

CourseCourseWorkDeleteCall
Deletes a course work.

CourseCourseWorkGetCall
Returns course work.

CourseCourseWorkListCall
Returns a list of course work that the requester is permitted to view.

CourseCourseWorkModifyAssigneeCall
Modifies assignee mode and options of a coursework.

CourseCourseWorkPatchCall
Updates one or more fields of a course work.

CourseCourseWorkStudentSubmissionGetCall
Returns a student submission.

CourseCourseWorkStudentSubmissionListCall
Returns a list of student submissions that the requester is permitted to view, factoring in the OAuth scopes of the request. - may be specified as the course_work_id to include student submissions for multiple course work items.

CourseCourseWorkStudentSubmissionModifyAttachmentCall
Modifies attachments of student submission.

CourseCourseWorkStudentSubmissionPatchCall
Updates one or more fields of a student submission.

CourseCourseWorkStudentSubmissionReclaimCall
Reclaims a student submission on behalf of the student that owns it.

CourseCourseWorkStudentSubmissionReturnCall
Returns a student submission.

CourseCourseWorkStudentSubmissionTurnInCall
Turns in a student submission.

CourseCreateCall
Creates a course.

CourseDeleteCall
Deletes a course.

CourseGetCall
Returns a course.

CourseListCall
Returns a list of courses that the requesting user is permitted to view, restricted to those that match the request. Returned courses are ordered by creation time, with the most recently created coming first.

CourseMaterial
A material attached to a course as part of a material set.

CourseMaterialSet
A set of materials that appears on the "About" page of the course. These materials might include a syllabus, schedule, or other background information relating to the course as a whole.

CourseMethods
A builder providing access to all methods supported on course resources. It is not used directly, but through the Classroom hub.

CoursePatchCall
Updates one or more fields in a course.

CourseRosterChangesInfo
Information about a Feed with a feed_type of COURSE_ROSTER_CHANGES.

CourseStudentCreateCall
Adds a user as a student of a course.

CourseStudentDeleteCall
Deletes a student of a course.

CourseStudentGetCall
Returns a student of a course.

CourseStudentListCall
Returns a list of students of this course that the requester is permitted to view.

CourseTeacherCreateCall
Creates a teacher of a course.

CourseTeacherDeleteCall
Deletes a teacher of a course.

CourseTeacherGetCall
Returns a teacher of a course.

CourseTeacherListCall
Returns a list of teachers of this course that the requester is permitted to view.

CourseTopicCreateCall
Creates a topic.

CourseTopicDeleteCall
Deletes a topic.

CourseTopicGetCall
Returns a topic.

CourseTopicListCall
Returns the list of topics that the requester is permitted to view.

CourseTopicPatchCall
Updates one or more fields of a topic.

CourseUpdateCall
Updates a course.

CourseWork
Course work created by a teacher for students of the course.

CourseWorkChangesInfo
Information about a Feed with a feed_type of COURSE_WORK_CHANGES.

Date
Represents a whole or partial calendar date, e.g. a birthday. The time of day and time zone are either specified elsewhere or are not significant. The date is relative to the Proleptic Gregorian Calendar. This can represent:

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DriveFile
Representation of a Google Drive file.

DriveFolder
Representation of a Google Drive folder.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Feed
A class of notifications that an application can register to receive. For example: "all roster changes for a domain".

Form
Google Forms item.

GlobalPermission
Global user permission description.

GradeHistory
The history of each grade on this submission.

Guardian
Association between a student and a guardian of that student. The guardian may receive information about the student's course work.

GuardianInvitation
An invitation to become the guardian of a specified user, sent to a specified email address.

IndividualStudentsOptions
Assignee details about a coursework/announcement. This field is set if and only if assigneeMode is INDIVIDUAL_STUDENTS.

Invitation
An invitation to join a course.

InvitationAcceptCall
Accepts an invitation, removing it and adding the invited user to the teachers or students (as appropriate) of the specified course. Only the invited user may accept an invitation.

InvitationCreateCall
Creates an invitation. Only one invitation for a user and course may exist at a time. Delete and re-create an invitation to make changes.

InvitationDeleteCall
Deletes an invitation.

InvitationGetCall
Returns an invitation.

InvitationListCall
Returns a list of invitations that the requesting user is permitted to view, restricted to those that match the list request.

InvitationMethods
A builder providing access to all methods supported on invitation resources. It is not used directly, but through the Classroom hub.

JsonServerError
A utility type which can decode a server response that indicates error

Link
URL item.

ListAnnouncementsResponse
Response when listing course work.

ListCourseAliasesResponse
Response when listing course aliases.

ListCourseWorkResponse
Response when listing course work.

ListCoursesResponse
Response when listing courses.

ListGuardianInvitationsResponse
Response when listing guardian invitations.

ListGuardiansResponse
Response when listing guardians.

ListInvitationsResponse
Response when listing invitations.

ListStudentSubmissionsResponse
Response when listing student submissions.

ListStudentsResponse
Response when listing students.

ListTeachersResponse
Response when listing teachers.

ListTopicResponse
Response when listing topics.

Material
Material attached to course work.

MethodInfo
Contains information about an API request.

ModifyAnnouncementAssigneesRequest
Request to modify assignee mode and options of an announcement.

ModifyAttachmentsRequest
Request to modify the attachments of a student submission.

ModifyCourseWorkAssigneesRequest
Request to modify assignee mode and options of a coursework.

ModifyIndividualStudentsOptions
Contains fields to add or remove students from a course work or announcement where the assigneeMode is set to INDIVIDUAL_STUDENTS.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

MultipleChoiceQuestion
Additional details for multiple-choice questions.

MultipleChoiceSubmission
Student work for a multiple-choice question.

Name
Details of the user's name.

RangeResponseHeader

ReclaimStudentSubmissionRequest
Request to reclaim a student submission.

Registration
An instruction to Classroom to send notifications from the feed to the provided destination.

RegistrationCreateCall
Creates a Registration, causing Classroom to start sending notifications from the provided feed to the destination provided in cloudPubSubTopic.

RegistrationDeleteCall
Deletes a Registration, causing Classroom to stop sending notifications for that Registration.

RegistrationMethods
A builder providing access to all methods supported on registration resources. It is not used directly, but through the Classroom hub.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ReturnStudentSubmissionRequest
Request to return a student submission.

ServerError

ServerMessage

SharedDriveFile
Drive file that is used as material for course work.

ShortAnswerSubmission
Student work for a short answer question.

StateHistory
The history of each state this submission has been in.

Student
Student in a course.

StudentSubmission
Student submission for course work.

SubmissionHistory
The history of the submission. This currently includes state and grade histories.

Teacher
Teacher of a course.

TimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

Topic
Topic created by a teacher for the course

TurnInStudentSubmissionRequest
Request to turn in a student submission.

UserProfile
Global information for a user.

UserProfileGetCall
Returns a user profile.

UserProfileGuardianDeleteCall
Deletes a guardian.

UserProfileGuardianGetCall
Returns a specific guardian.

UserProfileGuardianInvitationCreateCall
Creates a guardian invitation, and sends an email to the guardian asking them to confirm that they are the student's guardian.

UserProfileGuardianInvitationGetCall
Returns a specific guardian invitation.

UserProfileGuardianInvitationListCall
Returns a list of guardian invitations that the requesting user is permitted to view, filtered by the parameters provided.

UserProfileGuardianInvitationPatchCall
Modifies a guardian invitation.

UserProfileGuardianListCall
Returns a list of guardians that the requesting user is permitted to view, restricted to those that match the request.

UserProfileMethods
A builder providing access to all methods supported on userProfile resources. It is not used directly, but through the Classroom hub.

XUploadContentType
The X-Upload-Content-Type header.

YouTubeVideo
YouTube video item.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.