Struct google_classroom1::Course

pub struct Course {
    pub update_time: Option<String>,
    pub description: Option<String>,
    pub enrollment_code: Option<String>,
    pub course_group_email: Option<String>,
    pub course_material_sets: Option<Vec<CourseMaterialSet>>,
    pub course_state: Option<String>,
    pub id: Option<String>,
    pub room: Option<String>,
    pub name: Option<String>,
    pub alternate_link: Option<String>,
    pub section: Option<String>,
    pub creation_time: Option<String>,
    pub teacher_group_email: Option<String>,
    pub teacher_folder: Option<DriveFolder>,
    pub owner_id: Option<String>,
    pub description_heading: Option<String>,
}

A Course in Classroom.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

course work student submissions patch courses (none)

course work student submissions list courses (none)

get courses (response)

update courses (request|response)

students delete courses (none)

teachers get courses (none)

course work list courses (none)

teachers list courses (none)

course work student submissions turn in courses (none)

course work student submissions modify attachments courses (none)

course work student submissions return courses (none)

course work get courses (none)

course work create courses (none)

list courses (none)

course work student submissions reclaim courses (none)

aliases create courses (none)

students create courses (none)

aliases delete courses (none)

create courses (request|response)

students list courses (none)

delete courses (none)

patch courses (request|response)

aliases list courses (none)

teachers delete courses (none)

teachers create courses (none)

course work student submissions get courses (none)

students get courses (none)

Fields
update_time: Option<String>

Time of the most recent update to this course. Specifying this field in a course update mask results in an error. Read-only.

description: Option<String>

Optional description. For example, "We'll be learning about the structure of living creatures from a combination of textbooks, guest lectures, and lab work. Expect to be excited!" If set, this field must be a valid UTF-8 string and no longer than 30,000 characters.

enrollment_code: Option<String>

Enrollment code to use when joining this course. Specifying this field in a course update mask results in an error. Read-only.

course_group_email: Option<String>

The email address of a Google group containing all members of the course. This group does not accept email and can only be used for permissions. Read-only.

course_material_sets: Option<Vec<CourseMaterialSet>>

Sets of materials that appear on the "about" page of this course. Read-only.

course_state: Option<String>

State of the course. If unspecified, the default state is PROVISIONED.

id: Option<String>

Identifier for this course assigned by Classroom. When creating a course, you may optionally set this identifier to an alias string in the request to create a corresponding alias. The id is still assigned by Classroom and cannot be updated after the course is created. Specifying this field in a course update mask results in an error.

room: Option<String>

Optional room location. For example, "301". If set, this field must be a valid UTF-8 string and no longer than 650 characters.

name: Option<String>

Name of the course. For example, "10th Grade Biology". The name is required. It must be between 1 and 750 characters and a valid UTF-8 string.

alternate_link: Option<String>

Absolute link to this course in the Classroom web UI. Read-only.

section: Option<String>

Section of the course. For example, "Period 2". If set, this field must be a valid UTF-8 string and no longer than 2800 characters.

creation_time: Option<String>

Creation time of the course. Specifying this field in a course update mask results in an error. Read-only.

teacher_group_email: Option<String>

The email address of a Google group containing all teachers of the course. This group does not accept email and can only be used for permissions. Read-only.

teacher_folder: Option<DriveFolder>

Information about a Drive Folder that is shared with all teachers of the course. This field will only be set for teachers of the course and domain administrators. Read-only.

owner_id: Option<String>

The identifier of the owner of a course. When specified as a parameter of a create course request, this field is required. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user This must be set in a create request. Specifying this field in a course update mask results in an INVALID_ARGUMENT error.

description_heading: Option<String>

Optional heading for the description. For example, "Welcome to 10th Grade Biology." If set, this field must be a valid UTF-8 string and no longer than 3600 characters.