Struct google_classroom1::CourseMethods

pub struct CourseMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on course resources. It is not used directly, but through the Classroom hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_classroom1 as classroom1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use classroom1::Classroom;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Classroom::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `aliases_create(...)`, `aliases_delete(...)`, `aliases_list(...)`, `create(...)`, `delete(...)`, `get(...)`, `list(...)`, `patch(...)`, `students_create(...)`, `students_delete(...)`, `students_get(...)`, `students_list(...)`, `teachers_create(...)`, `teachers_delete(...)`, `teachers_get(...)`, `teachers_list(...)` and `update(...)`
// to build up your call.
let rb = hub.courses();

Methods

impl<'a, C, A> CourseMethods<'a, C, A>

fn teachers_list(&self, course_id: &str) -> CourseTeacherListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of teachers of this course that the requester is permitted to view. This method returns the following error codes: * NOT_FOUND if the course does not exist. * PERMISSION_DENIED for access errors.

Arguments

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

fn get(&self, id: &str) -> CourseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID.

Arguments

id - Identifier of the course to return. This identifier can be either the Classroom-assigned identifier or an alias.

fn students_delete(&self, course_id: &str, user_id: &str) -> CourseStudentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete students of this course or for access errors. * NOT_FOUND if no student of this course has the requested ID or if the course does not exist.

Arguments

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

userId - Identifier of the student to delete. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn teachers_get(&self, course_id: &str, user_id: &str) -> CourseTeacherGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view teachers of this course or for access errors. * NOT_FOUND if no teacher of this course has the requested ID or if the course does not exist.

Arguments

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

userId - Identifier of the teacher to return. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn aliases_list(&self, course_id: &str) -> CourseAliaseListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of aliases for a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to access the course or for access errors. * NOT_FOUND if the course does not exist.

Arguments

courseId - The identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

fn update(&self, request: Course, id: &str) -> CourseUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to modify the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID. * FAILED_PRECONDITION for the following request errors: * CourseNotModifiable

Arguments

request - No description provided.

id - Identifier of the course to update. This identifier can be either the Classroom-assigned identifier or an alias.

fn aliases_create(&self, request: CourseAlias, course_id: &str) -> CourseAliaseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an alias for a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create the alias or for access errors. * NOT_FOUND if the course does not exist. * ALREADY_EXISTS if the alias already exists.

Arguments

request - No description provided.

courseId - Identifier of the course to alias. This identifier can be either the Classroom-assigned identifier or an alias.

fn list(&self) -> CourseListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of courses that the requesting user is permitted to view, restricted to those that match the request. This method returns the following error codes: * PERMISSION_DENIED for access errors. * INVALID_ARGUMENT if the query argument is malformed. * NOT_FOUND if any users specified in the query arguments do not exist.

fn students_create(&self, request: Student, course_id: &str) -> CourseStudentCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a user as a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create students in this course or for access errors. * NOT_FOUND if the requested course ID does not exist. * FAILED_PRECONDITION if the requested user's account is disabled, for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * UserGroupsMembershipLimitReached * ALREADY_EXISTS if the user is already a student or teacher in the course.

Arguments

request - No description provided.

courseId - Identifier of the course to create the student in. This identifier can be either the Classroom-assigned identifier or an alias.

fn aliases_delete(&self, course_id: &str, alias: &str) -> CourseAliaseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an alias of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to remove the alias or for access errors. * NOT_FOUND if the alias does not exist.

Arguments

courseId - Identifier of the course whose alias should be deleted. This identifier can be either the Classroom-assigned identifier or an alias.

alias - Alias to delete. This may not be the Classroom-assigned identifier.

fn create(&self, request: Course) -> CourseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a course. The user specified in ownerId is the owner of the created course and added as a teacher. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create courses or for access errors. * NOT_FOUND if the primary teacher is not a valid user. * FAILED_PRECONDITION if the course owner's account is disabled or for the following request errors: * UserGroupsMembershipLimitReached * ALREADY_EXISTS if an alias was specified in the id and already exists.

Arguments

request - No description provided.

fn students_list(&self, course_id: &str) -> CourseStudentListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of students of this course that the requester is permitted to view. This method returns the following error codes: * NOT_FOUND if the course does not exist. * PERMISSION_DENIED for access errors.

Arguments

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

fn delete(&self, id: &str) -> CourseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID.

Arguments

id - Identifier of the course to delete. This identifier can be either the Classroom-assigned identifier or an alias.

fn patch(&self, request: Course, id: &str) -> CoursePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates one or more fields in a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to modify the requested course or for access errors. * NOT_FOUND if no course exists with the requested ID. * INVALID_ARGUMENT if invalid fields are specified in the update mask or if no update mask is supplied. * FAILED_PRECONDITION for the following request errors: * CourseNotModifiable

Arguments

request - No description provided.

id - Identifier of the course to update. This identifier can be either the Classroom-assigned identifier or an alias.

fn students_get(&self, course_id: &str, user_id: &str) -> CourseStudentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a student of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to view students of this course or for access errors. * NOT_FOUND if no student of this course has the requested ID or if the course does not exist.

Arguments

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

userId - Identifier of the student to return. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn teachers_delete(&self, course_id: &str, user_id: &str) -> CourseTeacherDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to delete teachers of this course or for access errors. * NOT_FOUND if no teacher of this course has the requested ID or if the course does not exist. * FAILED_PRECONDITION if the requested ID belongs to the primary teacher of this course.

Arguments

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.

userId - Identifier of the teacher to delete. The identifier can be one of the following: * the numeric identifier for the user * the email address of the user * the string literal "me", indicating the requesting user

fn teachers_create(&self, request: Teacher, course_id: &str) -> CourseTeacherCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a teacher of a course. This method returns the following error codes: * PERMISSION_DENIED if the requesting user is not permitted to create teachers in this course or for access errors. * NOT_FOUND if the requested course ID does not exist. * FAILED_PRECONDITION if the requested user's account is disabled, for the following request errors: * CourseMemberLimitReached * CourseNotModifiable * CourseTeacherLimitReached * UserGroupsMembershipLimitReached * ALREADY_EXISTS if the user is already a teacher or student in the course.

Arguments

request - No description provided.

courseId - Identifier of the course. This identifier can be either the Classroom-assigned identifier or an alias.