Struct google_classroom1::CourseWork

pub struct CourseWork {
    pub update_time: Option<String>,
    pub description: Option<String>,
    pub course_id: Option<String>,
    pub assignment: Option<Assignment>,
    pub associated_with_developer: Option<bool>,
    pub max_points: Option<f64>,
    pub id: Option<String>,
    pub work_type: Option<String>,
    pub submission_modification_mode: Option<String>,
    pub title: Option<String>,
    pub alternate_link: Option<String>,
    pub creation_time: Option<String>,
    pub due_date: Option<Date>,
    pub state: Option<String>,
    pub materials: Option<Vec<Material>>,
    pub due_time: Option<TimeOfDay>,
    pub multiple_choice_question: Option<MultipleChoiceQuestion>,
}

Course work created by a teacher for students of the course.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

course work create courses (request|response)

course work get courses (response)

Fields
update_time: Option<String>

Timestamp of the most recent change to this course work. Read-only.

description: Option<String>

Optional description of this course work. If set, the description must be a valid UTF-8 string containing no more than 30,000 characters.

course_id: Option<String>

Identifier of the course. Read-only.

assignment: Option<Assignment>

Assignment details. This is populated only when work_type is ASSIGNMENT.

associated_with_developer: Option<bool>

Whether this course work item is associated with the Developer Console project making the request. See google.classroom.Work.CreateCourseWork for more details. Read-only.

max_points: Option<f64>

Maximum grade for this course work. If zero or unspecified, this assignment is considered ungraded. This must be a non-negative integer value.

id: Option<String>

Classroom-assigned identifier of this course work, unique per course. Read-only.

work_type: Option<String>

Type of this course work. The type is set when the course work is created and cannot be changed. When creating course work, this must be ASSIGNMENT.

submission_modification_mode: Option<String>

Setting to determine when students are allowed to modify submissions. If unspecified, the default value is MODIFIABLE_UNTIL_TURNED_IN.

title: Option<String>

Title of this course work. The title must be a valid UTF-8 string containing between 1 and 3000 characters.

alternate_link: Option<String>

Absolute link to this course work in the Classroom web UI. This is only populated if state is PUBLISHED. Read-only.

creation_time: Option<String>

Timestamp when this course work was created. Read-only.

due_date: Option<Date>

Optional date, in UTC, that submissions for this this course work are due. This must be specified if due_time is specified.

state: Option<String>

Status of this course work. If unspecified, the default state is DRAFT.

materials: Option<Vec<Material>>

Additional materials. CourseWork must have no more than 20 material items.

due_time: Option<TimeOfDay>

Optional time of day, in UTC, that submissions for this this course work are due. This must be specified if due_date is specified.

multiple_choice_question: Option<MultipleChoiceQuestion>

Multiple choice question details. This is populated only when work_type is MULTIPLE_CHOICE_QUESTION.